Feature/data rendering for roles

+++ ScientificReport/Controllers/DepartmentController.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using ScientificReport.BLL.Interfaces;
@@ -13,7 +14,7 @@ using ScientificReport.DTO.Models.Department;
 
 namespace ScientificReport.Controllers
 {
-//	[Authorize(Roles = UserProfileRole.Administrator)]
+	[Authorize(Roles = UserProfileRole.Any)]
 	public class DepartmentController : Controller
 	{
 		private readonly UserManager<UserProfile> _userManager;
@@ -35,7 +36,21 @@ namespace ScientificReport.Controllers
 		[HttpGet]
 		public IActionResult Index()
 		{
-			return View(_departmentService.GetAll());
+			if (PageHelpers.IsAdmin(User))
+			{
+				return View(_departmentService.GetAll());
+			}
+
+			var department = _departmentService.Get(
+				d => d.Staff.Contains(_userProfileService.Get(User))
+			);
+
+			if (department == null)
+			{
+				return NotFound();
+			}
+
+			return RedirectToAction("Details", new { id = department.Id });
 		}
 
 		// GET: Department/Details/{id}
@@ -53,19 +68,31 @@ namespace ScientificReport.Controllers
 				return NotFound();
 			}
 
+			if (!PageHelpers.IsAdmin(User) && !department.Staff.Contains(_userProfileService.Get(User)))
+			{
+				return Forbid();
+			}
+
 			return View(department);
 		}
 
 		// GET: Department/Create
 		[HttpGet]
-		public IActionResult Create()
+		[Authorize(Roles = UserProfileRole.Administrator)]
+		public async Task<IActionResult> Create()
 		{
 			var allUsers = _userProfileService.GetAll();
 			var allDepartments = _departmentService.GetAll();
 
-			var availableUsers = allUsers
-				.Where(user => allDepartments.All(d => d.Head.Id != user.Id && !d.Staff.Contains(user)))
-				.Select(user => new SelectItem(user.FullName, user.Id.ToString()));
+			var usersToCheck = allUsers.Where(user => allDepartments.All(d => d.Head.Id != user.Id && !d.Staff.Contains(user))).ToList();
+			var availableUsers = new List<UserProfile>();
+			foreach (var user in usersToCheck)
+			{
+				if (await _userProfileService.IsTeacherOnlyAsync(user, _userManager))
+				{
+					availableUsers.Add(user);
+				}
+			}
 			
 			return View(new DepartmentCreateModel
 			{
@@ -75,7 +102,8 @@ namespace ScientificReport.Controllers
 
 		// POST: Department/Create
 		[HttpPost]
-		public IActionResult Create(DepartmentCreateModel model)
+		[Authorize(Roles = UserProfileRole.Administrator)]
+		public async Task<IActionResult> Create(DepartmentCreateModel model)
 		{
 			if (!ModelState.IsValid)
 			{
@@ -90,12 +118,26 @@ namespace ScientificReport.Controllers
 			if (_userProfileService.UserExists(model.SelectedHeadId))
 			{
 				var head = _userProfileService.GetById(model.SelectedHeadId);
-				_departmentService.CreateItem(new Department
+				var isAvailable =
+					!(await _userProfileService.IsInRoleAsync(head, UserProfileRole.Administrator, _userManager)
+					&& await _userProfileService.IsInRoleAsync(head, UserProfileRole.HeadOfDepartment, _userManager));
+
+				if (isAvailable)
 				{
-					Title = model.Title,
-					Head = head,
-					Staff = new List<UserProfile> {head}
-				});
+					_departmentService.CreateItem(new Department
+					{
+						Title = model.Title,
+						Head = head,
+						Staff = new List<UserProfile> {head}
+					});
+					head.Position = "HeadOfDepartment";
+					_userProfileService.UpdateItem(head);
+					await _userProfileService.AddToRoleAsync(head, UserProfileRole.HeadOfDepartment, _userManager);	
+				}
+				else
+				{
+					return BadRequest();
+				}
 			}
 			else
 			{
@@ -107,7 +149,7 @@ namespace ScientificReport.Controllers
 
 		// GET: Department/Edit/{id}
 		[HttpGet]
-//		[Authorize(Roles = UserProfileRole.HeadOfDepartmentOrAdmin)]
+		[Authorize(Roles = UserProfileRole.HeadOfDepartmentOrAdmin)]
 		public IActionResult Edit(Guid? id)
 		{
 			if (id == null)
@@ -116,40 +158,41 @@ namespace ScientificReport.Controllers
 			}
 
 			var department = _departmentService.GetById(id.Value);
+			if (department == null)
+			{
+				return RedirectToAction("Index");
+			}
+
+			if (!IsValidCurrentUser(department))
+			{
+				return Forbid();
+			}
+
 			var departments = _departmentService.GetAllWhere(d => d.Id != department.Id).ToList();
 			var allUsers = _userProfileService.GetAll();
 
-			var availableUsers = allUsers
-				.Where(user => departments.All(d => !d.Staff.Contains(user)))
-				.Select(user => new SelectItem(user.FullName, user.Id.ToString(), user.Id.Equals(department.Head.Id)));
+			var availableUsers = allUsers.Where(user => departments.All(d => !d.Staff.Contains(user)));
 
-			var availableScientificWorks = _scientificWorkService
-				.GetAllWhere(sw => departments.All(d => !d.ScientificWorks.Contains(sw)))
-				.Select(scientificWork => new SelectItem(scientificWork.Title, scientificWork.Id.ToString()));
+			var availableScientificWorks = _scientificWorkService.GetAllWhere(sw => departments.All(d => !d.ScientificWorks.Contains(sw)));
 
-			if (department != null)
+			return View(new DepartmentEditModel
 			{
-				return View(new DepartmentEditModel
-				{
-					DepartmentId = department.Id,
-					Title = department.Title,
-					UserSelection = availableUsers,
-					Head = department.Head,
-					SelectedHeadId = department.Head.Id,
-					Staff = department.Staff,
-					ScientificWorkItems = availableScientificWorks,
-					ScientificWorks = department.ScientificWorks,
-					// TODO: uncomment User.IsInRole... after 'login' is fixed
-					IsEditingByHead = true // User.IsInRole(UserProfileRole.Administrator)
-				});
-			}
-
-			return RedirectToAction("Index");
+				Department = department,
+				DepartmentId = department.Id,
+				Title = department.Title,
+				UserSelection = availableUsers,
+				Head = department.Head,
+				SelectedHeadId = department.Head.Id,
+				Staff = department.Staff,
+				AvailableScientificWork = availableScientificWorks,
+				ScientificWorks = department.ScientificWorks,
+				IsEditingByHead = department.Head.Id == _userProfileService.Get(u => u.UserName == User.Identity.Name).Id
+			});
 		}
 
 		// POST: Department/Edit/{id}
 		[HttpPost]
-//		[Authorize(Roles = UserProfileRole.HeadOfDepartmentOrAdmin)]
+		[Authorize(Roles = UserProfileRole.HeadOfDepartmentOrAdmin)]
 		public async Task<IActionResult> Edit(Guid? id, [Bind("Title,SelectedHeadId")] DepartmentEditModel model)
 		{
 			if (!ModelState.IsValid)
@@ -163,10 +206,14 @@ namespace ScientificReport.Controllers
 			}
 
 			var department = _departmentService.GetById(id.Value);
-
 			if (department == null)
 			{
 				return NotFound();
+			}
+
+			if (!IsValidCurrentUser(department))
+			{
+				return Forbid();
 			}
 
 			department.Title = model.Title;
@@ -175,16 +222,22 @@ namespace ScientificReport.Controllers
 			if (newHead != null && departments.All(d => !d.Staff.Contains(newHead)))
 			{
 				var oldHead = department.Head;
-				oldHead.Position = "Викладач";
+				oldHead.Position = "Teacher";
 				_userProfileService.UpdateItem(oldHead);
 				await _userProfileService.RemoveFromRoleAsync(oldHead, UserProfileRole.HeadOfDepartment,
 					_userManager);
 
 				department.Head = newHead;
 
-				newHead.Position = "Завідувач";
+				newHead.Position = "HeadOfDepartment";
 				_userProfileService.UpdateItem(newHead);
 				await _userProfileService.AddToRoleAsync(newHead, UserProfileRole.HeadOfDepartment, _userManager);
+
+				if (!department.Staff.Contains(newHead))
+				{
+					department.Staff.Add(newHead);
+					_departmentService.UpdateItem(department);
+				}
 			}
 
 			_departmentService.UpdateItem(department);
@@ -194,21 +247,32 @@ namespace ScientificReport.Controllers
 
 		// POST: Department/AddUserToStaff/{departmentId}
 		[HttpPost]
-		public IActionResult AddUserToStaff(Guid? id, [FromBody] DepartmentUpdateStaffRequest request)
+		[Authorize(Roles = UserProfileRole.HeadOfDepartmentOrAdmin)]
+		public async Task<IActionResult> AddUserToStaff(Guid? id, [FromBody] DepartmentUpdateStaffRequest request)
 		{
 			if (id == null)
 			{
 				return NotFound();
 			}
-
+			
 			var department = _departmentService.GetById(id.Value);
 			if (department == null)
+			{
 				return Json(ApiResponse.Fail);
+			}
+			
+			if (!IsValidCurrentUser(department))
+			{
+				return Json(ApiResponse.Fail);
+			}
 
 			var user = _userProfileService.GetById(request.UserId);
+			var isNotAvailable = await _userProfileService.IsInRoleAsync(user, UserProfileRole.HeadOfDepartment, _userManager);
 			
-			if (user == null || _departmentService.UserIsHired(user))
+			if (user == null || _departmentService.UserIsHired(user) || isNotAvailable)
+			{
 				return Json(ApiResponse.Fail);
+			}
 
 			department.Staff.Add(user);
 			_departmentService.UpdateItem(department);
@@ -218,6 +282,7 @@ namespace ScientificReport.Controllers
 
 		// POST: Department/RemoveUserFromStaff/{departmentId}
 		[HttpPost]
+		[Authorize(Roles = UserProfileRole.HeadOfDepartmentOrAdmin)]
 		public IActionResult RemoveUserFromStaff(Guid? id, [FromBody] DepartmentUpdateStaffRequest request)
 		{
 			if (id == null)
@@ -227,9 +292,16 @@ namespace ScientificReport.Controllers
 
 			var department = _departmentService.GetById(id.Value);
 			if (department == null)
+			{
 				return Json(ApiResponse.Fail);
-			var user = _userProfileService.GetById(request.UserId);
+			}
+
+			if (!IsValidCurrentUser(department))
+			{
+				return Json(ApiResponse.Fail);
+			}
 			
+			var user = _userProfileService.GetById(request.UserId);
 			if (user == null || !department.Staff.Contains(user) || department.Head.Id.Equals(user.Id))
 			{
 				return Json(ApiResponse.Fail);
@@ -243,6 +315,7 @@ namespace ScientificReport.Controllers
 
 		// POST: Department/AddScientificWork/{departmentId}
 		[HttpPost]
+		[Authorize(Roles = UserProfileRole.HeadOfDepartment)]
 		public IActionResult AddScientificWork(Guid? id, [FromBody] DepartmentUpdateScientificWorksRequest request)
 		{
 			if (id == null)
@@ -252,10 +325,20 @@ namespace ScientificReport.Controllers
 
 			var department = _departmentService.GetById(id.Value);
 			if (department == null)
+			{
 				return Json(new
 				{
 					Success = false
 				});
+			}
+
+			if (!IsValidCurrentUser(department))
+			{
+				return Json(new
+				{
+					Success = false
+				});
+			}
 			
 			var scientificWork = _scientificWorkService.GetById(request.ScientificWorkId);
 			if (scientificWork == null ||
@@ -278,6 +361,7 @@ namespace ScientificReport.Controllers
 
 		// POST: Department/RemoveScientificWork/{departmentId}
 		[HttpPost]
+		[Authorize(Roles = UserProfileRole.HeadOfDepartment)]
 		public IActionResult RemoveScientificWork(Guid? id, [FromBody] DepartmentUpdateScientificWorksRequest request)
 		{
 			if (id == null)
@@ -287,39 +371,63 @@ namespace ScientificReport.Controllers
 
 			var department = _departmentService.GetById(id.Value);
 			if (department == null)
-				return Json(new
-				{
-					Success = false
-				});
+			{
+				return Json(ApiResponse.Fail);
+			}
+
+			if (!IsValidCurrentUser(department))
+			{
+				return Json(ApiResponse.Fail);
+			}
+			
 			var scientificWork = _scientificWorkService.GetById(request.ScientificWorkId);
 			if (scientificWork == null || !department.ScientificWorks.Contains(scientificWork))
 			{
-				return Json(new
-				{
-					Success = false
-				});
+				return Json(ApiResponse.Fail);
 			}
 
 			department.ScientificWorks.Remove(scientificWork);
 			_departmentService.UpdateItem(department);
 
-			return Json(new
-			{
-				Success = true
-			});
+			return Json(ApiResponse.Ok);
 		}
 
 		// POST: Department/Delete/{id}
 		[HttpPost]
-		public IActionResult Delete(Guid? id)
+		[Authorize(Roles = UserProfileRole.Administrator)]
+		public async Task<IActionResult> Delete(Guid? id)
 		{
 			if (id == null || !_departmentService.Exists(id.Value))
 			{
 				return NotFound();
 			}
 
-			_departmentService.DeleteById(id.Value);
+			var departmentToDelete = _departmentService.GetById(id.Value);
+			if (!IsValidCurrentUser(departmentToDelete))
+			{
+				return Forbid();
+			}
+
+			if (departmentToDelete.Staff.Count == 1)
+			{
+				await _userProfileService.RemoveFromRoleAsync(departmentToDelete.Head, UserProfileRole.HeadOfDepartment, _userManager);
+				departmentToDelete.Head.Position = UserProfileRole.Teacher;
+				_userProfileService.UpdateItem(departmentToDelete.Head);
+				_departmentService.DeleteById(id.Value);
+			}
+
 			return RedirectToAction("Index");
+		}
+
+		private bool IsValidCurrentUser(Department department)
+		{
+			var currentUser = _userProfileService.Get(User);
+			if (!PageHelpers.IsAdmin(User))
+			{
+				return currentUser.Id == department.Head.Id;	
+			}
+
+			return true;
 		}
 	}
 }
