+++ ScientificReport/Controllers/UserProfileController.cs
@@ -1,17 +1,20 @@
 using System;
+using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.Extensions.Localization;
 using ScientificReport.BLL.Interfaces;
+using ScientificReport.Controllers.Utils;
 using ScientificReport.DAL.Entities.UserProfile;
 using ScientificReport.DAL.Roles;
 using ScientificReport.DTO.Models.UserProfile;
 
 namespace ScientificReport.Controllers
 {
-//	[Authorize(Roles = UserProfileRole.Administrator)]
+	[Authorize(Roles = UserProfileRole.Any)]
 	public class UserProfileController : Controller
 	{
 		private readonly UserManager<UserProfile> _userManager;
@@ -20,13 +23,15 @@ namespace ScientificReport.Controllers
 		
 		private readonly IUserProfileService _userProfileService;
 		private readonly IDepartmentService _departmentService;
+		private readonly IStringLocalizer<UserProfileController> _localizer;
 
 		public UserProfileController(
 			UserManager<UserProfile> usrMgr,
 			SignInManager<UserProfile> signInManager,
 			RoleManager<UserProfileRole> roleManager,
 			IUserProfileService userProfileService,
-			IDepartmentService departmentService
+			IDepartmentService departmentService,
+			IStringLocalizer<UserProfileController> localizer
 		)
 		{
 			_userManager = usrMgr;
@@ -34,18 +39,31 @@ namespace ScientificReport.Controllers
 			_roleManager = roleManager;
 			_userProfileService = userProfileService;
 			_departmentService = departmentService;
+			_localizer = localizer;
 		}
 
 		// GET: UserProfile/Index
 		[HttpGet]
+		[Authorize(Roles = UserProfileRole.HeadOfDepartmentOrAdmin)]
 		public IActionResult Index()
 		{
-			return View(_userProfileService.GetAll());
+			IEnumerable<UserProfile> users = null;
+			if (PageHelpers.IsAdmin(User))
+			{
+				users = _userProfileService.GetAll();
+			}
+			else if (PageHelpers.IsHeadOfDepartment(User))
+			{
+				var currentUser = _userProfileService.Get(User);
+				var department = _departmentService.Get(u => u.Head.Id == currentUser.Id);
+				users = department.Staff;
+			}
+			return View(users);
 		}
 
 		// GET: UserProfile/Details/{id}
 		[HttpGet]
-		public IActionResult Details(Guid? id)
+		public async Task<IActionResult> Details(Guid? id)
 		{
 			if (id == null)
 			{
@@ -59,12 +77,33 @@ namespace ScientificReport.Controllers
 			}
 
 			var department = _departmentService.Get(d => d.Staff.Contains(userProfile));
-			
+
+			var fullPositionTitle = "";
+			if (await _userManager.IsInRoleAsync(userProfile, UserProfileRole.Administrator))
+			{
+				fullPositionTitle = _localizer["Administrator"] + ", ";
+			}
+			fullPositionTitle += _localizer[userProfile.Position] + (department != null
+            					                    ? " " + _localizer["ofDepartment"] + " \"" + department.Title + "\""
+            					                    : "");
+
 			var detailsModel = new UserDetailsModel
 			{
 				User = userProfile,
-				DepartmentName = department != null ? "кафедри " + department.Title : ""
+				FullPositionTitle = fullPositionTitle
 			};
+
+			if (!PageHelpers.IsAdmin(User))
+			{
+				var currentUser = _userProfileService.Get(User);
+				if (PageHelpers.IsHeadOfDepartment(User))
+				{
+					if (department == null || !_departmentService.UserWorksInDepartment(currentUser, userProfile.Id))
+					{
+						return Forbid();
+					}
+				}	
+			}
 
 			return View(detailsModel);
 		}
@@ -79,6 +118,22 @@ namespace ScientificReport.Controllers
 			var user = _userProfileService.GetById(id.Value);
 			if (user != null)
 			{
+				var currentUser = _userProfileService.Get(User);
+				if (!PageHelpers.IsAdmin(User))
+				{
+					if (PageHelpers.IsHeadOfDepartment(User))
+					{
+						if (!_departmentService.UserWorksInDepartment(currentUser, user.Id))
+						{
+							return Forbid();
+						}
+					}
+					else if (PageHelpers.IsTeacher(User) && currentUser.Id != user.Id)
+					{
+						return Forbid();
+					}
+				}
+
 				return View(new UserProfileEditModel
 				{
 					UserId = user.Id,
@@ -93,13 +148,16 @@ namespace ScientificReport.Controllers
 					YearDegreeAssigned = user.YearDegreeAssigned,
 					PhoneNumber = user.PhoneNumber,
 					IsApproved = user.IsApproved,
+					IsActive = user.IsActive,
 					UserName = user.UserName,
 					Email = user.Email,
+					IsSelfEditing = currentUser.Id == user.Id,
 					AllRoles = _roleManager.Roles.ToList(),
-					UserRoles = await _userManager.GetRolesAsync(user)
+					UserRoles = await _userManager.GetRolesAsync(user),
+					IsHeadOfDepartment = await _userProfileService.IsInRoleAsync(user, UserProfileRole.HeadOfDepartment, _userManager)
 				});
 			}
-
+			
 			return RedirectToAction("Index");
 		}
 
@@ -120,33 +178,63 @@ namespace ScientificReport.Controllers
 			if (_userProfileService.UserExists(id.Value))
 			{
 				var user = _userProfileService.GetById(id.Value);
+				var currentUser = _userProfileService.Get(User);
+				if (PageHelpers.IsAdmin(User) || PageHelpers.IsTeacher(User) && currentUser.Id == user.Id || PageHelpers.IsHeadOfDepartment(User) && _departmentService.UserWorksInDepartment(currentUser, user.Id))
+				{
+					user.FirstName = model.FirstName;
+					user.MiddleName = model.MiddleName;
+					user.LastName = model.LastName;
 
-				user.FirstName = model.FirstName;
-				user.MiddleName = model.MiddleName;
-				user.LastName = model.LastName;
-				user.BirthYear = model.BirthYear;
-				user.GraduationYear = model.GraduationYear;
-				user.ScientificDegree = model.ScientificDegree;
-				user.YearDegreeGained = model.YearDegreeGained;
-				user.AcademicStatus = model.AcademicStatus;
-				user.YearDegreeAssigned = model.YearDegreeAssigned;
-				user.PhoneNumber = model.PhoneNumber;
-				user.IsApproved = model.IsApproved;
-				user.UserName = model.UserName;
-				user.Email = model.Email;
+					if (model.BirthYear > 1900)
+					{
+						user.BirthYear = model.BirthYear;	
+					}
+
+					if (model.GraduationYear > 1900)
+					{
+						user.GraduationYear = model.GraduationYear;	
+					}
+
+					if (model.ScientificDegree != null)
+					{
+						user.ScientificDegree = model.ScientificDegree;
+						user.YearDegreeGained = model.YearDegreeGained;	
+					}
+
+					if (model.AcademicStatus != null)
+					{
+						user.AcademicStatus = model.AcademicStatus;
+						user.YearDegreeAssigned = model.YearDegreeAssigned;	
+					}
+					
+					user.PhoneNumber = model.PhoneNumber;
+					if (PageHelpers.IsAdminOrHead(User) && currentUser.Id != id.Value)
+					{
+						user.IsApproved = model.IsApproved;
+						user.IsActive = model.IsActive;
+					}
+					
+					user.UserName = model.UserName;
+					user.Email = model.Email;
 				
-				_userProfileService.UpdateItem(user);
+					_userProfileService.UpdateItem(user);
+				}
+				else
+				{
+					return Forbid();
+				}
 			}
 			else
 			{
 				return NotFound();
 			}
-			
-			return RedirectToAction("Index");
+
+			return PageHelpers.IsAdminOrHead(User) ? RedirectToAction("Index") : RedirectToAction("Details", "UserProfile", new { id = id.Value});
 		}
 
 		// POST: UserProfile/AddUserToAdministration/{userId}
 		[HttpPost]
+		[Authorize(Roles = UserProfileRole.Administrator)]
 		public async Task<IActionResult> AddUserToAdministration(Guid? id, [FromBody] UserProfileUpdateRolesRequest request)
 		{
 			if (id == null)
@@ -164,10 +252,15 @@ namespace ScientificReport.Controllers
 			if (userExists)
 			{
 				var user = _userProfileService.GetById(id.Value);
+				if (await _userProfileService.IsInRoleAsync(user, UserProfileRole.HeadOfDepartment, _userManager))
+				{
+					return Json(ApiResponse.Fail);
+				}
+				
 				if (!await _userProfileService.IsInRoleAsync(user, request.RoleName, _userManager))
 				{
 					await _userProfileService.AddToRoleAsync(user, request.RoleName, _userManager);
-				}	
+				}
 			}
 
 			return Json(new {Success = userExists});
@@ -175,6 +268,7 @@ namespace ScientificReport.Controllers
 		
 		// POST: UserProfile/RemoveUserFromAdministration/{userId}
 		[HttpPost]
+		[Authorize(Roles = UserProfileRole.Administrator)]
 		public async Task<IActionResult> RemoveUserFromAdministration(Guid? id, [FromBody] UserProfileUpdateRolesRequest request)
 		{
 			if (id == null)
@@ -188,12 +282,22 @@ namespace ScientificReport.Controllers
 				var user = _userProfileService.GetById(id.Value);
 				if (await _userProfileService.IsInRoleAsync(user, request.RoleName, _userManager))
 				{
-					await _userProfileService.RemoveFromRoleAsync(user, request.RoleName, _userManager);
-					if (request.RoleName == UserProfileRole.HeadOfDepartment)
+					if (user.UserName != User.Identity.Name)
 					{
-						user.Position = "Викладач";
-						_userProfileService.UpdateItem(user);
+						await _userProfileService.RemoveFromRoleAsync(user, request.RoleName, _userManager);
+						if (request.RoleName == UserProfileRole.HeadOfDepartment)
+						{
+							_userProfileService.UpdateItem(user);
+						}	
 					}
+					else
+					{
+						return Json(ApiResponse.Fail);
+					}
+				}
+				else
+				{
+					return Json(ApiResponse.Fail);
 				}
 			}
 			
@@ -205,6 +309,7 @@ namespace ScientificReport.Controllers
 
 		// POST: UserProfile/Delete/{id}
 		[HttpPost]
+		[Authorize(Roles = UserProfileRole.Administrator)]
 		public IActionResult Delete(Guid? id)
 		{
 			if (id == null)
@@ -217,14 +322,66 @@ namespace ScientificReport.Controllers
 				return NotFound();
 			}
 			
-			_userProfileService.DeleteById(id.Value);
+			var currentUser = _userProfileService.Get(User);
+			if (currentUser.Id != id.Value)
+			{
+				if (!PageHelpers.IsAdmin(User) && PageHelpers.IsHeadOfDepartment(User))
+				{
+					if (!_departmentService.UserWorksInDepartment(currentUser, id.Value))
+					{
+						return Forbid();
+					}
+				}
+				
+				_userProfileService.DeleteById(id.Value);	 
+			}
+			return RedirectToAction("Index");
+		}
+		
+		// POST: UserProfile/SetActive/{id}?active={true/false}
+		[HttpPost]
+		[Authorize(Roles = UserProfileRole.HeadOfDepartmentOrAdmin)]
+		public IActionResult SetActive(Guid? id, bool isActive)
+		{
+			if (id == null)
+			{
+				return NotFound();
+			}
+			
+			if (!_userProfileService.UserExists(id.Value))
+			{
+				return NotFound();
+			}
+			
+			var currentUser = _userProfileService.Get(User);
+			if (currentUser.Id != id.Value)
+			{
+				if (!PageHelpers.IsAdmin(User) && PageHelpers.IsHeadOfDepartment(User))
+				{
+					if (!_departmentService.UserWorksInDepartment(currentUser, id.Value))
+					{
+						return Forbid();
+					}
+				}
+				
+				_userProfileService.SetActiveById(id.Value, isActive);	 
+			}
+			
 			return RedirectToAction("Index");
 		}
 
 		// GET: UserProfile/Register
 		[HttpGet]
 		[AllowAnonymous]
-		public IActionResult Register() => View();
+		public IActionResult Register()
+		{
+			// .Select(d => new SelectItem(d.Title, d.Id.ToString()))
+			var departments = _departmentService.GetAll();
+			return View(new RegisterModel
+			{
+				Departments = departments
+			});
+		}
 		
 		// POST: UserProfile/Register
 		[HttpPost]
@@ -238,7 +395,7 @@ namespace ScientificReport.Controllers
 
 			if (_userProfileService.Get(usr => usr.UserName == model.UserName) != null)
 			{
-				ModelState.AddModelError(string.Empty, "User already exists");
+				ModelState.AddModelError(string.Empty, _localizer["UserAlreadyExists"]);
 				return BadRequest();
 			}
 			
@@ -248,14 +405,9 @@ namespace ScientificReport.Controllers
 				FirstName = model.FirstName,
 				LastName = model.LastName,
 				MiddleName = model.MiddleName,
-				BirthYear = model.BirthYear,
-				AcademicStatus = model.AcademicStatus,
-				GraduationYear = model.GraduationYear,
-				ScientificDegree = model.ScientificDegree,
-				YearDegreeGained = model.YearDegreeGained,
-				YearDegreeAssigned = model.YearDegreeAssigned,
-				Position = "Викладач",
+				Position = UserProfileRole.Teacher,
 				IsApproved = false,
+				IsActive = true,
 				PhoneNumber = model.PhoneNumber
 			};
 			if (model.Password.Equals(model.PasswordRepeat))
@@ -269,7 +421,13 @@ namespace ScientificReport.Controllers
 					);
 					if (addUserToRoleResult.Succeeded)
 					{
-						return RedirectToAction("Index");	
+						var department = _departmentService.GetById(model.SelectedDepartmentId);
+						if (department != null)
+						{
+							department.Staff.Add(createdUser);
+							_departmentService.UpdateItem(department);
+							return RedirectToAction("Index");
+						}
 					}
 					
 					AddErrorsFromResult(addUserToRoleResult);
@@ -279,7 +437,7 @@ namespace ScientificReport.Controllers
 			}
 			else
 			{
-				ModelState.AddModelError(string.Empty, "Password confirmation failed");
+				ModelState.AddModelError(string.Empty, _localizer["PasswordConfirmationFailed"]);
 			}
 			
 			return View(model);
@@ -309,7 +467,13 @@ namespace ScientificReport.Controllers
 				var user = _userProfileService.Get(usr => usr.UserName == model.UserName);
 				if (user != null)
 				{
-					if (user.IsApproved)
+					if (!user.IsApproved)
+					{
+						ModelState.AddModelError(string.Empty, _localizer["AccountNotApproved"]);
+						return View(model);	
+					}
+					
+					if (user.IsActive)
 					{
 						var result = await _signInManager.PasswordSignInAsync(
 							user.UserName, model.Password, model.RememberMe, false
@@ -319,14 +483,10 @@ namespace ScientificReport.Controllers
 							return Redirect(model.ReturnUrl);
 						}	
 					}
-					else
-					{
-						ModelState.AddModelError(string.Empty, "Account is not approved yet");
-						return View(model);
-					}
 				}
 			}
-			ModelState.AddModelError(string.Empty, "Invalid login or password");
+
+			ModelState.AddModelError(string.Empty, _localizer["IncorrectLoginOrPassword"]);
 			return View(model);
 		}
